Track 02: real held-out evaluation, and the detector it exposed as broken
The bar says 'measured precision and recall on a held-out test set'. The old
numbers were 24 hand-written cases that I had ALSO tuned the thresholds against,
reporting precision 1.00 / FPR 0.00. On a proper held-out split that same
detector scored precision 0.278 and FPR 12.1%. The earlier figures were fiction.

- engine/src/dataset.cpp: seeded, documented behaviour model, 400 sessions and
  8,719 transactions, including hard negatives (late lunch, team order,
  occasional merchant) so the normal class is not trivially separable. Split
  ~60/40 by SESSION hash, since a session's history is what the detector scores
  against and splitting mid-session would leak the answer. Full labelled set
  written to CSV for inspection.
- Tuning sweeps 192 combinations on TRAIN only; headline numbers are TEST.
  Train and test are printed side by side so overfitting is visible.

The root cause was design, not tuning: firing on ANY single signal made
R_NEW_MERCHANT an accusation, and 906 legitimate transactions in the set were at
a merchant the agent had not used before. Signals are now weighted (burst 3,
velocity 2, new merchant 1, odd hour 1, flag at >=3) and a never-seen merchant
only counts once the agent has an established narrow pattern.

  held-out: precision 0.908, recall 0.301, FPR 0.002 (0.20% of good traffic)
  train F1 0.401 vs test 0.452 -- no meaningful overfitting

Recall 0.30 is reported as the honest number with the full tradeoff curve
(recall 0.80 is available at 8.7% friction), plus why it is acceptable: this
detector sits BEHIND the deterministic caps, so a miss is still bounded.

Two harness bugs found on the way, both of which would have produced misleading
numbers: a failed threshold sweep silently reported untuned defaults as chosen,
and a printf format mismatch read a pointer as a double and segfaulted. The
build had no warning flags -- now -Wall -Wextra -Wformat=2, zero warnings.

Track 01: added /.well-known/agent-commerce and /api/catalog so an AI buyer can
discover the merchant without a human, which is what 'transactable end to end'
requires.

+++ engine/src/gateway.cpp
@@ -465,7 +465,7 @@ std::string Gateway::decision_json(const Decision& d) const {
 
   o << ",\"reasons\":[";
   bool first = true;
-  for (int b = 0; b < R_BIT_COUNT; ++b) {
+  for (int b = 0; b < static_cast<int>(R_BIT_COUNT); ++b) {
     const std::uint32_t bit = 1u << b;
     if (!(d.verdict.bits & bit)) continue;
     if (!first) o << ",";
